integrate ssi_preprocessing into features.processing
